+++ tests/testthat/test-5-deltaX_get.R
@@ -13,7 +13,7 @@ test_that("deltaX_get", {
   tt <- SEMseeker:::get_meth_tech(signal_data)
 
   ####################################################################################
-  sliding_window_size <- 11
+  lesion_window_kbp <- 5
   bonferroni_threshold <- 0.05
 
   if (!exists("signal_thresholds"))